Refactor private functions to allow testing (#10)
* Add findOne and fix some tests

* Refactor find functions

* Change back to jsldb :(

* Correct the docs according to the current spec

* remove references : findOne -> find1

* 0.0.9

* Change the stable version number to 1.0.0

I may have misunderstood the semantic version numbering when I first
started this project, or I may have forgotten the reasoning behind the
old number. Version 1.0.0 will the stable version.

* nuke the docs.

* nuke the docs.

* Salvaged from jsdoc

* Lots of changes

* refactor everything
* create new docs
* added `updateById` and its test
* probably other stuff

This is a pretty messy commit, sorry.

* Add a docs index

* Remove now-pointless .jsdoc.json

* 0.0.10

* Add a _private for testing

* Another huge commit, added _private for testing

I added a bunch of new tests and squashed a bunch of bugs.
I started a new test file for the private functions.

* Add a few more tests

* 0.0.11

+++ test/relational.public.test.js
@@ -80,8 +80,8 @@ describe('Creation, saving, and connection', () => {
         }
     })
 
-    test('Save (sync)', () => {
-        expect(db.saveSync()).toBe(true)
+    test('saveSync', () => {
+        expect(db.saveSync()).toBeTruthy()
     })
 
     test('Connect to existing database', () => {
@@ -131,32 +131,34 @@ describe('Entry insertion and modification', () => {
     })
 
     describe('Entry type check', () => {
-        test('Insert a broken entry (number)', (done) => {
+        test('Insert a broken entry (number)', () => {
             const entry = {
                 number: 'string',
                 string: 'string',
                 date: new Date(),
                 idLink: 'id table1'
             }
-            db.insert('table1', entry, (err, entry) => {
-                expect(err).toBeDefined()
-                expect(entry).toBeNull()
-                done()
-            })
+            expect(() => {
+                db.insert('table1', entry, (err, entry) => {
+                    expect(err).toBeDefined()
+                    expect(entry).toBeNull()
+                })
+            }).toThrow()
         })
 
-        test('Insert a broken entry (string)', (done) => {
+        test('Insert a broken entry (string)', () => {
             const entry = {
                 number: 42,
                 string: 42,
                 date: new Date(),
                 idLink: 'id table1'
             }
-            db.insert('table1', entry, (err, entry) => {
-                expect(err).toBeDefined()
-                expect(entry).toBeNull()
-                done()
-            })
+            expect(() => {
+                db.insert('table1', entry, (err, entry) => {
+                    expect(err).toBeDefined()
+                    expect(entry).toBeNull()
+                })
+            }).toThrow()
         })
 
         test('Insert a broken entry (date)', (done) => {
@@ -166,28 +168,26 @@ describe('Entry insertion and modification', () => {
                 date: 0,
                 idLink: 'id table1'
             }
-            db.insert('table1', entry, (err, entry) => {
-                expect(err).toBeDefined()
-                expect(entry).toBeNull()
-                done()
-            })
+            expect(() => {
+                db.insert('table1', entry, (err, entry) => {
+                    expect(err).toBeDefined()
+                    expect(entry).toBeNull()
+                    done()
+                })
+            }).toThrow()
         })
 
-        test('Insert broken entry (id)', (done) => {
+        test('Insert broken entry (id)', () => {
             const entry = {
                 number: 42,
                 string: 'string',
                 date: new Date(),
                 idLink: 'id table3'
             }
-            db.insert('table1', entry, (err, entry) => {
-                expect(err).toBeDefined()
-                expect(entry).toBeNull()
-                done()
-            })
+            expect(() => {
+                db.insert('table1', entry)
+            }).toThrow()
         })
-
-        // TODO: test failure of 'array id nonExistentTable'
     })
 
     test('Insert a table2 entry', (done) => {
@@ -244,6 +244,14 @@ describe('Entry insertion and modification', () => {
             expect(fs.existsSync(tPath)).toBe(true)
             done()
         })
+    })
+})
+
+describe('insertion errors', () => {
+    test('callback defined but not function type should throw', () => {
+        expect(() => {
+            db.insert('table', t1Entry, 'notafunction')
+        }).toThrow()
     })
 })
 
@@ -648,9 +656,7 @@ describe('Faker', () => {
 
 afterAll(cleanup)
 
-process.on('exit', cleanup)
-
 function cleanup () {
-    fs.unlinkSync(tPath)
-    fs.unlinkSync(`${tPath}.old`)
+    if (fs.existsSync(tPath)) fs.unlinkSync(tPath)
+    if (fs.existsSync(`${tPath}.old`)) fs.unlinkSync(`${tPath}.old`)
 }
